Update "about" text
Added new "about" text block, with the exception of the table which did not contain complete information.   We will need to revert the info modal to contain only the first paragraph under "About This Map" with a "Learn More" button, which will go to a full about page with the entire text.